+++ .gitignore
@@ -1,3 +1,8 @@
 .vscode
 downloads
 json
+__pycache__
+*.json
+*.mp4
+*.webp
+*.png
